Initial check-in
OK, we're going to try this once more.

This is a new check-in. It contains all of the work of the previous git
repository, but with a working Default.unity scene and no UI work.

This includes database and access,  NPC movement, Player movement,
models, crew portraits, and skyboxes. It does not include the crew UI
for the main HUD, however. That was lost in the Great Purge.