Station Main Menu Stub
* When inside the station trigger, the button will light up. Users can
click it with the mouse or use the F key to open the main menu
* Station menu manager was added. Handles tracking all station menus
that are open, and responding to messages to open/close station menus
from other locations
* Need to disable player movement until the station menus are all
closed. Also, might want to somehow hide or otherwise stealth the
player, so they don't get shot at or knocked around while in menus.