Commodity grid and Esc button
* Fixed the null reference issue. The prefab didn't have the script
being referenced, even though the original primitive in the scene did.
Had to add it to the prefab, and it fixed the issue
* Added a method to close windows in reverse order when the ESC button
is pressed. Had an issue when openWindowList.count wasn't taking the
zero-ness of the list into consideration, but then once the ESC button
was pressed, it closed the top-most window (last into the bucket).
* Turned off the navigation trail option on the "Commodity Menu" button
on the main station menu because with that turned on, it took the
spacebar to mean "click this button", so using space to re-activate the
ship nav clicked the button, re-opening the commodity menu.