Better Turning for the Cops
* Using some code found on the Internet, I added some nice
rotation-over-time to the police. It's not yet perfect, but they no
longer turn on a time. Have to adjust the point at which they start
turning (sooner), and how long it takes them to turn (slower).
* Also have to remove the VS integration dlls because for some reason
the different environments don't agree on when they should exist and
when they shouldn't