Legacy Data Migration
* Created a one-shot editor script that loaded the data from the
MasterDatabase into ScriptableObject databases.

I wanted to check this in at this point because it marks the Point of No
Return, where I will be heavily modifying the
DataController/MasterDatabase to work with the ScriptableObject system.
This may actually do away with the MD portion, allowing me to re-work
the DataController so it loads the data directly into lists accessible
via the singleton pattern.

Still, none of this deals with the dynamic, save game data, so that'll
need to be at the back of my mind as I work on this shift.