Switchover to ScriptableObjects Complete (Probably)
The switch from class-based to ScriptableObject based master data should
be all set.
* Created all of the SO database objects for each List<T> represented in
the MasterDatabase file, except for the NPCs which is a system that is
not yet complete.
* Created a legacy data translation editor. It has no buttons or
anything, but because SO's need to be handled by the Editor system, I
had to include it this way. It takes the List<T> from MasterDatabase and
creates instances in each of the SO databases for each original List<T>
* Created a SaveGameController. This is where the shop inventory lists
are stored, with a pass-through to DataController so users can query the
DC rather than having to jump between SGC and DC depending on what data
is needed
* Initial commodity shop population has been moved to the
MarketController script because it's a market thing. MarketController is
not a singleton, and must be instantiated (not sure if it'll always be
this way)

Next, I want to create an editor for reviewing the commodity assignment,
which will probably mean an editor for Stations, with fold-outs for the
different "places" in the station like Commodity Market, Shipyard, Bar,
etc. where data can be edited. Hopefully that'll work like I assume it
will.