Station Editor and Jumpgate Editor
* Completed the station editor for the time being. It allows for the
creation and editing of stations and the commodity market, which is the
only child dataset currently available
* Started the jumpgate editor which follows the commodity editor in
form. However, the jumpgate editor has a lot more ID references which
would benefit greatly from a popup list of sectors and/or other
jumpgates rather than forcing the admin to work with raw ID values. So
this is a check-in before I rip it apart to get additional data and try
to work a popup into this panel.