Jumpgates be Jumpin'
* Jumpgates have a slick interface!
- Select the home sector, destination sector, and destination sector
gate analogue from drop-downs!
* Had to heavily modify the Jumpgate Data Object and the
dbJumpdateDataItem classes to include the destinationGateID and
jumpgateName fields. I realized that players would need to be dropping
into a sector at a specific point based on which gate they used to get
there, so I needed a destinationGateID. Then in order to make assigning
that gate easier, I had to be able to identify the gate by something
other than a number, so jumpgateName got added.

Ideally, we'd have a node editor-like interface which would create a
node for each system we add (and could be used to create and edit
systems as well!). Each line that came off of that node would represent
a gate connection to another node. This would basically present the
whole game world as an interconnected map with visual linkages and
everything!

But for now, this works.