Pathfinding
* Trying a few different solutions. No off the shelf option seems to
work well with my weird setup, so I am having to roll my own.

I might need to create a custom object that encapsulates the sector,
each gate, and each distance to every other sector it's directly linked
to. I'm having a difficult time getting the distances to work so I don't
accidentally end up in a cul-de-sac, but I think that's mostly a LINQ
issue (and me not knowing how to use it efficiently)