Pathfinding
* Fixed the system so it calculates distance from possible next sectors
to the destination, not the current sector to possible next sectors. It
still takes the shortest route.
* However, this can still lead to dead ends or sectors which don't lead
to the destination. For that, all sectors need to connect to other
sectors so that there's no dead-ends. This was done as a test for one
sector, but other sectors need similar updates
* Deal with DENEB. This fucker has been a PITA from day one. He's not
properly wired according to the diagram, and needs to be re-worked so he
does.